+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#841](https://github.com/ClementTsang/bottom/pull/841): Add page up/page down support for the help screen.
 - [#868](https://github.com/ClementTsang/bottom/pull/868): Make temperature widget sortable.
 - [#870](https://github.com/ClementTsang/bottom/pull/870): Make disk widget sortable.
+- [#881](https://github.com/ClementTsang/bottom/pull/881): Add pasting to the search bar.
 
 ## [0.6.8] - 2022-02-01